# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BlackDiamondClub Gambling Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get current user wallet balances
      responses:
        "200":
          description: Wallet info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wallet/deposit:
    post:
      operationId: createDeposit
      tags: [wallet]
      summary: Submit a USDT BEP20 deposit request with screenshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "200":
          description: Deposit processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /wallet/withdraw:
    post:
      operationId: createWithdrawal
      tags: [wallet]
      summary: Request a withdrawal from withdrawal wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalInput"
      responses:
        "200":
          description: Withdrawal submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Insufficient funds or below minimum
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List user transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [deposit, withdrawal, game_win, game_loss]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Wallet:
      type: object
      properties:
        userId:
          type: integer
        depositBalance:
          type: number
        withdrawalBalance:
          type: number
        currency:
          type: string
      required:
        - userId
        - depositBalance
        - withdrawalBalance
        - currency
...
    DepositInput:
      type: object
      properties:
        amount:
          type: number
        txHash:
          type: ["string", "null"]
        screenshotBase64:
          type: ["string", "null"]
          description: Base64 encoded screenshot image of the payment
      required:
        - amount
    WithdrawalInput:
      type: object
      properties:
        amount:
          type: number
        walletAddress:
          type: string
      required:
        - amount
        - walletAddress
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, game_win, game_loss]
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        txHash:
          type: ["string", "null"]
        screenshotUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        gameType:
          type: ["string", "null"]
      required:
        - id
        - userId
        - type
        - amount
        - status
        - createdAt
    TransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
      required:
        - transactions
        - total
...